requirements.txt
json
os
sys
transformers
torch
faiss-cpu
sentence-transformers
numpy
sentence_transformers
datasets

